feat: Refactor teams test suites
