+++ apps/web/playwright/teams/managed-event-types.e2e.ts
@@ -1,6 +1,6 @@
 import { expect } from "@playwright/test";
 
-import { test } from "./lib/fixtures";
+import { test } from "../lib/fixtures";
 
 test.afterEach(({ users }) => users.deleteAll());
